+++ hugr-passes/src/dead_funcs.rs
@@ -3,16 +3,16 @@
 use std::collections::HashSet;
 
 use hugr_core::{
-    HugrView, Node,
+    HugrView, Node, Visibility,
     hugr::hugrmut::HugrMut,
     module_graph::{ModuleGraph, StaticNode},
     ops::{OpTag, OpTrait},
 };
+use itertools::Either;
 use petgraph::visit::{Dfs, Walker};
 
-use crate::{
-    ComposablePass,
-    composable::{ValidatePassError, validate_if_test},
+use crate::composable::{
+    ComposablePass, PassScope, Preserve, ValidatePassError, WithScope, validate_if_test,
 };
 
 #[derive(Debug, thiserror::Error)]
@@ -48,15 +48,26 @@ fn reachable_funcs<'a, H: HugrView>(
     })
 }
 
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 /// A configuration for the Dead Function Removal pass.
 pub struct RemoveDeadFuncsPass {
-    entry_points: Vec<Node>,
+    entry_points: Either<Vec<Node>, PassScope>,
+}
+
+impl Default for RemoveDeadFuncsPass {
+    fn default() -> Self {
+        Self {
+            entry_points: Either::Left(Vec::new()),
+        }
+    }
 }
 
 impl RemoveDeadFuncsPass {
+    #[deprecated(note = "Use RemoveDeadFuncsPass::with_scope", since = "0.25.7")]
     /// Adds new entry points - these must be [`FuncDefn`] nodes
     /// that are children of the [`Module`] at the root of the Hugr.
+    ///
+    /// Overrides any [PassScope] set by a call to [Self::with_scope].
     ///
     /// [`FuncDefn`]: hugr_core::ops::OpType::FuncDefn
     /// [`Module`]: hugr_core::ops::OpType::Module
@@ -64,7 +75,14 @@ impl RemoveDeadFuncsPass {
         mut self,
         entry_points: impl IntoIterator<Item = Node>,
     ) -> Self {
-        self.entry_points.extend(entry_points);
+        let v = match self.entry_points {
+            Either::Left(ref mut v) => v,
+            Either::Right(_) => {
+                self.entry_points = Either::Left(Vec::new());
+                self.entry_points.as_mut().unwrap_left()
+            }
+        };
+        v.extend(entry_points);
         self
     }
 }
@@ -72,17 +90,44 @@ impl RemoveDeadFuncsPass {
 impl<H: HugrMut<Node = Node>> ComposablePass<H> for RemoveDeadFuncsPass {
     type Error = RemoveDeadFuncsError;
     type Result = ();
+
     fn run(&self, hugr: &mut H) -> Result<(), RemoveDeadFuncsError> {
         let mut entry_points = Vec::new();
-        for &n in self.entry_points.iter() {
-            if !hugr.get_optype(n).is_func_defn() {
-                return Err(RemoveDeadFuncsError::InvalidEntryPoint { node: n });
+        match &self.entry_points {
+            Either::Left(ep) => {
+                for &n in ep {
+                    if !hugr.get_optype(n).is_func_defn() {
+                        return Err(RemoveDeadFuncsError::InvalidEntryPoint { node: n });
+                    }
+                    debug_assert_eq!(hugr.get_parent(n), Some(hugr.module_root()));
+                    entry_points.push(n);
+                }
+                if hugr.entrypoint() != hugr.module_root() {
+                    entry_points.push(hugr.entrypoint())
+                }
             }
-            debug_assert_eq!(hugr.get_parent(n), Some(hugr.module_root()));
-            entry_points.push(n);
-        }
-        if hugr.entrypoint() != hugr.module_root() {
-            entry_points.push(hugr.entrypoint())
+            Either::Right(
+                // If the entrypoint is the module root, not allowed to touch anything.
+                // Otherwise, we must keep the entrypoint (and can touch only inside it).
+                PassScope::EntrypointFlat | PassScope::EntrypointRecursive
+                // Optimize whole Hugr but keep all functions
+                | PassScope::Global(Preserve::All)) => {
+                return Ok(());
+            }
+            Either::Right(PassScope::Global(Preserve::Entrypoint)) if hugr.entrypoint() != hugr.module_root() => {
+                entry_points.push(hugr.entrypoint());
+            }
+            Either::Right(PassScope::Global(_)) => {
+                for n in hugr.children(hugr.module_root()) {
+                    if hugr.get_optype(n).as_func_defn().is_some_and(|fd| fd.visibility() == &Visibility::Public)
+                    {
+                        entry_points.push(n);
+                    }
+                }
+                if hugr.entrypoint() != hugr.module_root() {
+                    entry_points.push(hugr.entrypoint());
+                }
+            }
         }
 
         let mut reachable =
@@ -109,8 +154,15 @@ impl<H: HugrMut<Node = Node>> ComposablePass<H> for RemoveDeadFuncsPass {
     }
 }
 
-/// Deletes from the Hugr any functions that are not used by either [`Call`] or
-/// [`LoadFunction`] nodes in reachable parts.
+impl WithScope for RemoveDeadFuncsPass {
+    fn with_scope(mut self, scope: impl Into<PassScope>) -> Self {
+        self.entry_points = Either::Right(scope.into());
+        self
+    }
+}
+
+/// Deletes from the Hugr any functions that are not used by either `Call` or
+/// `LoadFunction` nodes in reachable parts.
 ///
 /// `entry_points` may provide a list of entry points, which must be [`FuncDefn`]s (children of the root).
 /// The [HugrView::entrypoint] will also be used unless it is the [HugrView::module_root].
@@ -120,10 +172,10 @@ impl<H: HugrMut<Node = Node>> ComposablePass<H> for RemoveDeadFuncsPass {
 /// # Errors
 /// * If any node in `entry_points` is not a [`FuncDefn`]
 ///
-/// [`Call`]: hugr_core::ops::OpType::Call
 /// [`FuncDefn`]: hugr_core::ops::OpType::FuncDefn
-/// [`LoadFunction`]: hugr_core::ops::OpType::LoadFunction
 /// [`Module`]: hugr_core::ops::OpType::Module
+#[deprecated(note = "Use RemoveDeadFuncsPass with a PassScope", since = "0.25.7")]
+#[expect(deprecated)]
 pub fn remove_dead_funcs(
     h: &mut impl HugrMut<Node = Node>,
     entry_points: impl IntoIterator<Item = Node>,
@@ -138,52 +190,72 @@ pub fn remove_dead_funcs(
 mod test {
     use std::collections::HashMap;
 
+    use hugr_core::builder::{Dataflow, DataflowSubContainer, HugrBuilder, ModuleBuilder};
+    use hugr_core::hugr::hugrmut::HugrMut;
     use hugr_core::ops::handle::NodeHandle;
+    use hugr_core::{Hugr, Visibility};
+    use hugr_core::{HugrView, extension::prelude::usize_t, types::Signature};
     use itertools::Itertools;
     use rstest::rstest;
 
-    use hugr_core::builder::{Dataflow, DataflowSubContainer, HugrBuilder, ModuleBuilder};
-    use hugr_core::hugr::hugrmut::HugrMut;
-    use hugr_core::{HugrView, extension::prelude::usize_t, types::Signature};
+    use super::RemoveDeadFuncsPass;
+    use crate::PassScope;
+    use crate::composable::{Preserve, WithScope, test::run_validating};
 
-    use super::remove_dead_funcs;
+    fn hugr(use_entrypoint: bool) -> Hugr {
+        let mut hb = ModuleBuilder::new();
+        let o2 = hb
+            .define_function("from_pub", Signature::new_endo([usize_t()]))
+            .unwrap();
+        let o2inp = o2.input_wires();
+        let o2 = o2.finish_with_outputs(o2inp).unwrap();
+        let mut o1 = hb
+            .define_function_vis(
+                "pubfunc",
+                Signature::new_endo([usize_t()]),
+                Visibility::Public,
+            )
+            .unwrap();
+
+        let o1c = o1.call(o2.handle(), &[], o1.input_wires()).unwrap();
+        o1.finish_with_outputs(o1c.outputs()).unwrap();
+
+        let fm = hb
+            .define_function("from_main", Signature::new_endo([usize_t()]))
+            .unwrap();
+        let f_inp = fm.input_wires();
+        let fm = fm.finish_with_outputs(f_inp).unwrap();
+        let mut m = hb
+            .define_function("main", Signature::new_endo([usize_t()]))
+            .unwrap();
+        let m_in = m.input_wires();
+        let mut dfb = m
+            .dfg_builder(Signature::new_endo([usize_t()]), m_in)
+            .unwrap();
+        let c = dfb.call(fm.handle(), &[], dfb.input_wires()).unwrap();
+        let dfg = dfb.finish_with_outputs(c.outputs()).unwrap();
+        m.finish_with_outputs(dfg.outputs()).unwrap();
+        let mut h = hb.finish_hugr().unwrap();
+        if use_entrypoint {
+            h.set_entrypoint(dfg.node());
+        }
+        h
+    }
 
     #[rstest]
     #[case(false, [], vec![])] // No entry_points removes everything!
     #[case(true, [], vec!["from_main", "main"])]
     #[case(false, ["main"], vec!["from_main", "main"])]
     #[case(false, ["from_main"], vec!["from_main"])]
-    #[case(false, ["other1"], vec!["other1", "other2"])]
-    #[case(true, ["other2"], vec!["from_main", "main", "other2"])]
-    #[case(false, ["other1", "other2"], vec!["other1", "other2"])]
+    #[case(false, ["pubfunc"], vec!["from_pub", "pubfunc"])]
+    #[case(true, ["from_pub"], vec!["from_main", "from_pub", "main"])]
+    #[case(false, ["from_pub", "pubfunc"], vec!["from_pub", "pubfunc"])]
     fn remove_dead_funcs_entry_points(
         #[case] use_hugr_entrypoint: bool,
         #[case] entry_points: impl IntoIterator<Item = &'static str>,
         #[case] retained_funcs: Vec<&'static str>,
     ) -> Result<(), Box<dyn std::error::Error>> {
-        let mut hb = ModuleBuilder::new();
-        let o2 = hb.define_function("other2", Signature::new_endo([usize_t()]))?;
-        let o2inp = o2.input_wires();
-        let o2 = o2.finish_with_outputs(o2inp)?;
-        let mut o1 = hb.define_function("other1", Signature::new_endo([usize_t()]))?;
-
-        let o1c = o1.call(o2.handle(), &[], o1.input_wires())?;
-        o1.finish_with_outputs(o1c.outputs())?;
-
-        let fm = hb.define_function("from_main", Signature::new_endo([usize_t()]))?;
-        let f_inp = fm.input_wires();
-        let fm = fm.finish_with_outputs(f_inp)?;
-        let mut m = hb.define_function("main", Signature::new_endo([usize_t()]))?;
-        let m_in = m.input_wires();
-        let mut dfg = m.dfg_builder(Signature::new_endo([usize_t()]), m_in)?;
-        let c = dfg.call(fm.handle(), &[], dfg.input_wires())?;
-        let dfg = dfg.finish_with_outputs(c.outputs()).unwrap();
-        m.finish_with_outputs(dfg.outputs())?;
-
-        let mut hugr = hb.finish_hugr()?;
-        if use_hugr_entrypoint {
-            hugr.set_entrypoint(dfg.node());
-        }
+        let mut hugr = hugr(use_hugr_entrypoint);
 
         let avail_funcs = hugr
             .children(hugr.module_root())
@@ -194,7 +266,8 @@ mod test {
             })
             .collect::<HashMap<_, _>>();
 
-        remove_dead_funcs(
+        #[expect(deprecated)]
+        super::remove_dead_funcs(
             &mut hugr,
             entry_points
                 .into_iter()
@@ -214,5 +287,33 @@ mod test {
             .collect_vec();
         assert_eq!(remaining_funcs, retained_funcs);
         Ok(())
+    }
+
+    #[rstest]
+    #[case(Preserve::All, false, vec!["from_main", "from_pub", "main", "pubfunc"])]
+    #[case(PassScope::EntrypointFlat, true, vec!["from_main", "from_pub", "main", "pubfunc"])]
+    #[case(PassScope::EntrypointRecursive, false, vec!["from_main", "from_pub", "main", "pubfunc"])]
+    #[case(Preserve::Public, true, vec!["from_main", "from_pub", "main", "pubfunc"])]
+    #[case(Preserve::Public, false, vec!["from_pub", "pubfunc"])]
+    #[case(Preserve::Entrypoint, true, vec!["from_main", "main"])]
+    fn remove_dead_funcs_scope(
+        #[case] scope: impl Into<PassScope>,
+        #[case] use_entrypoint: bool,
+        #[case] retained_funcs: Vec<&'static str>,
+    ) {
+        let scope = scope.into();
+        let mut hugr = hugr(use_entrypoint);
+        run_validating(RemoveDeadFuncsPass::default().with_scope(scope), &mut hugr).unwrap();
+
+        let remaining_funcs = hugr
+            .nodes()
+            .filter_map(|n| {
+                hugr.get_optype(n)
+                    .as_func_defn()
+                    .map(|fd| fd.func_name().as_str())
+            })
+            .sorted()
+            .collect_vec();
+        assert_eq!(remaining_funcs, retained_funcs);
     }
 }