+++ test_files/hugr_examples/generate_simple_hugr.py
@@ -1,0 +1,148 @@
+"""A list of functions generating simple Hugr programs for testing"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+import hugr.ops as ops
+import hugr.tys as tys
+import hugr.val as val
+from hugr.build.dfg import Dfg
+from hugr.build.function import Module
+from hugr.std.int import INT_T, DivMod
+from hugr.std.logic import Not
+from sys import argv
+
+
+if TYPE_CHECKING:
+    from hugr.hugr import Hugr
+
+direct_call = [True, False]
+
+
+def simple_hugr() -> Hugr:
+    d = Dfg(tys.Qubit, tys.Bool, INT_T, INT_T)
+    d.metadata["name"] = "simple_hugr"
+    q, b, i1, i2 = d.inputs()
+
+    b = d.add_op(Not, b, metadata={"name": "not"})
+    c = d.add_op(Not, d.load(val.TRUE), metadata={"name": "not_true"})
+    i1, i2 = d.add(DivMod(i1, i2))
+    d.set_outputs(q, b, b, c, i1, i2)
+
+    return d.hugr
+
+
+(Path(argv[0]).parent / "simple_hugr.hugr").write_bytes(simple_hugr().to_bytes())
+
+
+def hugr_nested() -> Hugr:
+    h = Dfg(tys.Bool)
+    (a,) = h.inputs()
+
+    with h.add_nested(a) as nested:
+        (a1,) = nested.inputs()
+        nt = nested.add(Not(a1))
+        nested.set_outputs(nt)
+
+    h.set_outputs(nested)
+
+    return h.hugr
+
+
+(Path(argv[0]).parent / "hugr_nested.hugr").write_bytes(hugr_nested().to_bytes())
+
+
+def hugr_nested_w_external_edge() -> Hugr:
+    h = Dfg(tys.Bool, tys.Bool)
+    (a, b) = h.inputs()
+    with h.add_nested() as nested:
+        nt = nested.add(Not(a=a))
+        nested.set_outputs(nt)
+
+    h.set_outputs(nested, b)
+
+    return h.hugr
+
+
+(Path(argv[0]).parent / "hugr_nested_w_external_edge.hugr").write_bytes(
+    hugr_nested_w_external_edge().to_bytes()
+)
+
+
+def hugr_w_function_call(direct_call: bool) -> Hugr:
+    mod = Module()
+
+    f_id = mod.declare_function(
+        "literals",
+        tys.PolyFuncType(
+            [
+                tys.StringParam(),
+                tys.BoundedNatParam(),
+                tys.BytesParam(),
+                tys.FloatParam(),
+            ],
+            tys.FunctionType.endo([tys.Qubit]),
+        ),
+    )
+
+    f_main = mod.define_main([tys.Qubit])
+    q = f_main.inputs()[0]
+    # for now concrete instantiations have to be provided.
+    instantiation = tys.FunctionType.endo([tys.Qubit])
+    type_args = [
+        tys.StringArg("string"),
+        tys.BoundedNatArg(42),
+        tys.BytesArg(b"HUGR"),
+        tys.FloatArg(0.9),
+    ]
+    if direct_call:
+        call = f_main.call(f_id, q, instantiation=instantiation, type_args=type_args)
+    else:
+        load = f_main.load_function(
+            f_id, instantiation=instantiation, type_args=type_args
+        )
+        call = f_main.add(ops.CallIndirect()(load, q))
+
+    f_main.set_outputs(call)
+
+    return mod.hugr
+
+
+for dc in direct_call:
+    (Path(argv[0]).parent / f"hugr_w_function_{"direct" if dc else "indirect"}_call.hugr").write_bytes(
+        hugr_w_function_call(dc).to_bytes()
+    )
+
+
+def hugr_w_recursive_function() -> Hugr:
+    mod = Module()
+
+    f_recursive = mod.define_function("recurse", [tys.Qubit])
+    f_recursive.declare_outputs([tys.Qubit])
+    call = f_recursive.call(f_recursive, f_recursive.input_node[0])
+    f_recursive.set_outputs(call)
+
+    return mod.hugr
+
+
+(Path(argv[0]).parent / "hugr_w_recursive_function.hugr").write_bytes(
+    hugr_w_recursive_function().to_bytes()
+)
+
+
+def hugr_conditional() -> Hugr:
+    dfg = Dfg(tys.Tuple(tys.Bool, tys.Bool))
+    bool1, _ = dfg.add_op(ops.UnpackTuple(), *dfg.inputs())
+    cond = dfg.add_conditional(bool1)
+    with cond.add_case(0) as case:
+        case.set_outputs(bool1)
+    with cond.add_case(1) as case:
+        case.set_outputs(bool1)
+    dfg.set_outputs(*cond.outputs())
+
+    return dfg.hugr
+
+
+(Path(argv[0]).parent / "hugr_conditional.hugr").write_bytes(hugr_conditional().to_bytes())